Added conversion rates to 'Send ADA' screen
